Refactor: routes
