--- src/main/java/devkor/com/teamcback/domain/navigate/dto/response/DeleteNodeRes.java
@@ -1,10 +1,0 @@
-package devkor.com.teamcback.domain.navigate.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(description = "노드 삭제 응답 dto")
-@JsonIgnoreProperties
-public class DeleteNodeRes {
-
-}
