--- src/main/java/devkor/com/teamcback/domain/navigate/entity/Checkpoint.java
@@ -1,28 +1,0 @@
-package devkor.com.teamcback.domain.navigate.entity;
-
-import devkor.com.teamcback.domain.common.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "tb_checkpoint_name")
-@NoArgsConstructor
-public class Checkpoint extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne
-    @JoinColumn(name = "node_id")
-    private Node node;
-
-    private String name;
-}
