--- src/main/java/devkor/com/teamcback/domain/navigate/entity/Edge.java
@@ -1,19 +1,0 @@
-package devkor.com.teamcback.domain.navigate.entity;
-
-import devkor.com.teamcback.domain.common.BaseEntity;
-import lombok.Getter;
-
-@Getter
-public class Edge extends BaseEntity {
-    private int distance;
-
-    private Long startNode;
-
-    private Long endNode;
-
-    public Edge(int distance, Long startNode, Long endNode) {
-        this.distance = distance;
-        this.startNode = startNode;
-        this.endNode = endNode;
-    }
-}
