--- src/main/java/devkor/com/teamcback/domain/navigate/entity/LocationType.java
@@ -1,11 +1,0 @@
-package devkor.com.teamcback.domain.navigate.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum LocationType {
-    COORD, BUILDING, CLASSROOM, FACILITY;
-
-}
