+++ src/main/java/devkor/com/teamcback/domain/place/service/AdminPlaceService.java
@@ -11,8 +11,8 @@ import devkor.com.teamcback.domain.place.dto.response.GetPlaceRes;
 import devkor.com.teamcback.domain.place.dto.response.ModifyPlaceRes;
 import devkor.com.teamcback.domain.place.entity.Place;
 import devkor.com.teamcback.domain.place.repository.PlaceRepository;
-import devkor.com.teamcback.domain.navigate.entity.Node;
-import devkor.com.teamcback.domain.navigate.repository.NodeRepository;
+import devkor.com.teamcback.domain.routes.entity.Node;
+import devkor.com.teamcback.domain.routes.repository.NodeRepository;
 import devkor.com.teamcback.global.exception.GlobalException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
