+++ src/main/java/devkor/com/teamcback/domain/routes/entity/LocationType.java
@@ -1,0 +1,11 @@
+package devkor.com.teamcback.domain.routes.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum LocationType {
+    COORD, BUILDING, PLACE, NODE;
+
+}
